HealthGeneratorApp/pyproject.toml
[project]
name = "healthgeneratorapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Amin Benarieb",email = "amin.benarieb@welltory.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pillow = "^11.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
